Fixed space member properties not being shown
