+++ components/common/UserProfile/components/NftsList.tsx
@@ -49,7 +49,7 @@ export function NftsList({ userId, readOnly = false, isFetchingNfts, mutateNfts,
           <LoadingComponent isLoading />
         ) : (
           <Stack gap={2} display='flex' flexDirection='row'>
-            {pinnedNfts.length === 0 ? (
+            {pinnedNfts.length === 0 && readOnly ? (
               <Typography color='secondary'>No pinned NFTs</Typography>
             ) : (
               pinnedNfts
