+++ components/members/MemberDirectoryPageWithContext.tsx
@@ -1,10 +1,5 @@
 import MemberDirectoryPage from 'components/members/MemberDirectoryPage';
-import { MemberPropertiesProvider } from 'hooks/useMemberProperties';
 
 export default function MemberDirectoryPageWithContext() {
-  return (
-    <MemberPropertiesProvider>
-      <MemberDirectoryPage />
-    </MemberPropertiesProvider>
-  );
+  return <MemberDirectoryPage />;
 }
